Browser Updates
- Consolidate DriverOptions and BrowserOptions into BrowserConfiguration.
- Movie BrowserType enum to its own file.
- Remove unnecessary parameters from Browser constructors.
- Add BrowserConfiguration to IBrowser interface.
