+++ Bromine/Bromine.csproj
@@ -50,14 +50,14 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Constants\BrowserType.cs" />
     <Compile Include="Core\Browser.cs" />
     <Compile Include="Core\Driver.cs" />
     <Compile Include="Core\Element.cs" />
     <Compile Include="Core\Find.cs" />
     <Compile Include="Core\Navigate.cs" />
     <Compile Include="IBrowser.cs" />
-    <Compile Include="Models\BrowserOptions.cs" />
-    <Compile Include="Models\DriverOptions.cs" />
+    <Compile Include="Models\BrowserConfiguration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
