+++ Bromine/Constants/BrowserType.cs
@@ -1,0 +1,12 @@
+﻿namespace Bromine.Constants
+{
+    /// <summary>
+    /// Supported web browser types.
+    /// </summary>
+    public enum BrowserType
+    {
+        Chrome,
+        Firefox,
+        Edge
+    }
+}
